% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report-module.R
\name{reportMainUI}
\alias{reportMainUI}
\title{UI for the Report in the Main Panel}
\usage{
reportMainUI(id)
}
\arguments{
\item{id}{The module id}
}
\value{
A Shiny UI element
}
\description{
UI for the Report in the Main Panel
}
