% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report-module.R
\name{reportServer}
\alias{reportServer}
\title{Server for the Report Module}
\usage{
reportServer(
  id,
  bullet_data = NULL,
  comp_bul1 = NULL,
  comp_bul2 = NULL,
  phase_test_results = NULL
)
}
\arguments{
\item{id}{The module id}

\item{bullet_data}{A data frame}

\item{comp_bul1}{The name of the first selected bullet}

\item{comp_bul2}{The name of the second selected bullet}

\item{phase_test_results}{The results of the phase test}
}
\description{
Server for the Report Module
}
