% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report-module.R
\name{reportSidebarUI}
\alias{reportSidebarUI}
\title{UI for the Report Sidebar}
\usage{
reportSidebarUI(id)
}
\arguments{
\item{id}{The module id}
}
\value{
A Shiny UI element
}
\description{
UI for the Report Sidebar
}
