% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{bulletAnalyzrApp}
\alias{bulletAnalyzrApp}
\title{BulletAnalyzr Application}
\usage{
bulletAnalyzrApp(run_interactive = TRUE, ...)
}
\arguments{
\item{run_interactive}{TRUE allows the user to adjust the crosscut and groove
locations. FALSE uses the crosscut and grooves automatically chosen by the
app.}

\item{...}{Other arguments passed on to 'onStart', 'options', 'uiPattern', or
'enableBookmarking' of 'shiny::shinyApp'}
}
\value{
No return value, called to launch 'shiny' app

A Shiny app
}
\description{
Lauch a 'shiny' application for forensic bullet comparisons.
}
\examples{
\dontrun{
  bulletAnalyzrApp()
}

}
\keyword{Shiny}
