% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{bulletAnalyzrApp}
\alias{bulletAnalyzrApp}
\title{BulletAnalyzr Application}
\usage{
bulletAnalyzrApp(...)
}
\arguments{
\item{...}{Optional arguments passed to shiny::shinyApp}
}
\value{
No return value, called to launch 'shiny' app

A Shiny app
}
\description{
Lauch a 'shiny' application for forensic bullet comparisons.
}
\examples{
\dontrun{
bulletAnalyzrApp()
}

}
\keyword{Shiny}
